+++ CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## [TBA]
+### Added
+- Log in the info level, the request and response XMLs if logging is enabled
+
 ## [0.2.0] - 2016-03-09
 ### Added
 - `read_timeout` and `open_timeout` to set the timeout values, in seconds, to the SOAP client